requirements.txt
Flask==2.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2
fpdf2==2.7.4
joblib==1.3.2